A lock-free (when no APCs) algorithm to soothe Wlad.
The approach is epoch-based. Each time we enter and exit the task, we
increase the epoch. This way, we can tell either we are in the task
processing or we are finished and standing by, or going to, by checking
epoch oddity. Besides, if the task ran several times, we can know that apc
requests have been guaranteedly processed.

However, if the last apc request processing happens before exit epoch
increase, we cannot tell, whether it's been processed or not, and should
wait in the busy loop until either the task runs again (epoch changes),
or we remove fd from poll.

To prevent long busy loop, apc request processing should happen right
before start_io() and after epoch exit increase.

We can still end up busy-waiting until all the requests will be processed.
But since the request queue is guarded by LOCK_thd_kill, we'll end up in a
deadlock.

To prevent this, another epoch counter is introduced inside apc.
It increases once before the mutex lock (LOCK_thd_kill), and before it
unlocks in the end.

So by checking the oddity of this new counter we can tell, whether are we
hanging on the lock, or not. It is safe to consider that the request will
be processed if the epoch is odd.

This way, we