MDEV-28075: JSON_VALUE returns first value from array not from range
Analysis: When searching for the given path in json string, if the current
step is of array range type, then path was considered reached. So always
the first index of range is returned (starting index for '*')
Fix: If the current step type for a path is array range, then check if the
value in array_counter[] is in range of n_item and n_item_end. If this is true
then path is reached and only then return the value.